refactor: create custom hooks folder
