+++ frontend/components/auth/Login.tsx
@@ -6,7 +6,7 @@ import { Button } from '../ui/button'
 import { InputField } from '../customs/custom-input'
 import { PasswordReset } from './PasswordReset'
 import { toast } from 'sonner'
-import usePasswordToggle from '../account/UsePasswordToggle'
+import usePasswordToggle from '../../hooks/UsePasswordToggle'
 import { useState } from 'react'
 
 export default function Login() {
