+++ frontend/components/auth/PasswordReset.tsx
@@ -15,7 +15,7 @@ import validateInput, { initialFormValues } from '@/util/input-validation'
 import { Button } from '@/components/ui/button'
 import { ResetPasswordSteps } from '@/types/reset-password'
 import { toast } from 'sonner'
-import usePasswordToggle from '../account/UsePasswordToggle'
+import usePasswordToggle from '../../hooks/UsePasswordToggle'
 import { useState } from 'react'
 
 export function PasswordReset() {
